Create LevelRepository and repo-driven UI
Lay the foundation for the repository driven ui for each level. Most static content for a given level can now be defined by a LevelConfig. The new StaticLevelStage composable also allows for the game stage (static content) to be either repository-driven or make use of a custom composable for a particular level. Adjustments were made to the ViewModel to allow for state to be loaded in via the repository. The game screen was modified to make use of the new UI state.