Connect user input to sprite movement for Columns
Lays groundwork for the user to type into a textfield and move the Capybara when modifying either verticalArrangement or horizontalAlignment properties within a Column layout. Other layouts (as specified by the AnswerType enum) will be added next. Care could also be taken to avoid repetition via a sealed class or something similar in the future.