Allow for two input fields in CodeField.
Adjust UI and viewmodel to accommodate for two ui fields. Pattern in ViewModel's isUserInputCorrect() and GameScreen can be followed to extend to as many inputs in the code field as desired. Current level designs will likely need up to four inputs.

Some work still required to show when the input is correct. However, more logic is required to update the container additively to allow for more than one parameter dictating the ui container. This may also change the validation logic for the isCorrect to be displayed to the user.