Fix answer validation to work with 2 input fields
Answers are now validated by checking if the capybaraStageLayout container is equal to the StaticStageLayout container. This allows for properties any order of alignment and arrangement.

Answer validation will need to be modified again for modifiers but this is a step in the right direction to allow for more levels.