Fix web and iOS coroutine bug
There was an issue with my attempts at optimization using jobs where the user input would update the isCorrect state properly on iOS and web. A user would need to type the right answer, then delete it within 500ms to move on to the next stage. This has been resolved.